parking_platform/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
pillow>=10.0.0
torch>=2.0.0
modelscope>=1.9.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
pydeck>=0.8.0

